seperate logic from display
